+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Add support to Laravel package auto-discovery
+- Add support for Laravel package auto-discovery
 - Remove register facade alias
 
 ## 1.2.0 (2017-06-26)